% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_rules.R
\name{count_rules}
\alias{count_rules}
\title{Count total rules}
\usage{
count_rules(rules)
}
\arguments{
\item{rules}{Rule object}
}
\value{
Integer count
}
\description{
Counts all variable-level rules in a rule object.
}
\keyword{internal}
