% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rules.R
\name{load_rules}
\alias{load_rules}
\title{Title: Load yaml files}
\usage{
load_rules(file, validate = TRUE, verbose = TRUE)
}
\arguments{
\item{validate}{logical: whether to validate the structure of the rules file}

\item{verbose}{logical: whether to print messages about the loading process}

\item{file:}{structured yaml file containing rules}
}
\value{
A list of rules with class "datavalid_rules" and an attribute "source" containing the file path
}
\description{
Title: Load yaml files
}
\examples{
#' # Assuming you have a rules.yaml file structured correctly
rules <- load_rules("rules.yaml")

}
