% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null_coalesce}
\alias{null_coalesce}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First value}

\item{y}{Fallback value}
}
\value{
One of the two values
}
\description{
Returns the left-hand value if it is not NULL, otherwise returns the right-hand value.
}
\keyword{internal}
