% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_rules.R
\name{print.abayflowr_rules}
\alias{print.abayflowr_rules}
\title{Print abayflowr rules}
\usage{
\method{print}{abayflowr_rules}(x, ...)
}
\arguments{
\item{x}{An object of class "abayflowr_rules"}

\item{...}{Not used}
}
\description{
Provides a concise summary when a rules object is printed.
}
