% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_rule_structure.R
\name{validate_rule_structure}
\alias{validate_rule_structure}
\title{Validate the structure of YAML rules}
\usage{
validate_rule_structure(rules)
}
\arguments{
\item{rules}{A list read from YAML}
}
\value{
The same rules list if valid
}
\description{
Ensures that the rules YAML has the required sections and correct format.
}
\examples{
rules <- list(
  variables = list(
    age = list(
      type = "numeric",
      rules = list(list(rule = "min", value = 0))
    )
  )
)
validate_rule_structure(rules)
}
