% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rules.R
\name{load_rules}
\alias{load_rules}
\title{Load validation rules from a YAML file}
\usage{
load_rules(file, validate = TRUE, verbose = TRUE)
}
\arguments{
\item{file}{Path to YAML file containing rules}

\item{validate}{Logical. Validate the structure of the rules file}

\item{verbose}{Logical. Print loading messages}
}
\value{
A list of rules with class "abayflowr_rules" and attribute "source"
}
\description{
Load validation rules from a YAML file
}
\examples{
\dontrun{
rules <- load_rules("rules.yaml")
}
}
